refactor(launcher): split app/worker along view/state + messages/impl seams
app.rs (894) -> app/mod.rs (state struct, lifecycle, pure helpers + tests)
+ app/view.rs (eframe::App impl + show_* panels).

worker.rs (756) -> worker/mod.rs (Worker impl, spawn_*, upload task + tests)
+ worker/messages.rs (Command / Event / LaunchTelemetryConfig wire types).

Behavior-preserving pure code movement. Re-export discipline keeps
crate::worker::{Command,Event,LaunchTelemetryConfig} and crate::app::LauncherApp
valid; zero public-surface change.

Part of #529.
